add interpreter tests for loops in new format
Input files can now contain a block with variable type and name and
expected value. The variables in this block will be checked in the
global environment of the sketch.
The first line filled with '>' (the minimum is three) will act as a
separator between the code to test and the variables to check.